OSJ-168: On-disk caching of dynamically acquired metadata
Added fix for subtle issue that: If the metadata filter mutates the XMLObject,
then that will cause DOM drop and issues with persistence to cache if the
XMLObject can't be rountripped, e.g. signatures. So in that case filter on
a clone of the root object input.  This would also be relevant even when we
have the planned origin byte[] object metadata, since that will be dropped
along with the DOM.